requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
mstarpy
streamlit-local-storage
scipy
plotly
Riskfolio-Lib
pytest
Pyrebase4
playwright
dotenv
psycopg2-binary
